+++ lib/image_composition.dart
@@ -39,6 +39,8 @@ class _Composed {
 }
 
 /// The [ImageComposition] allows for composing multiple images onto a single image.
+///
+/// **Note:** Composing images is a heavy async operation and should not be called on each [Game.render].
 class ImageComposition {
   /// The values that will be used to compose the image
   final List<_Composed> _composes = [];