+++ CHANGELOG.md
@@ -20,7 +20,8 @@
  - Fix `SpriteAnimationComponent.shouldRemove` use `Component.shouldRemove`
  - Add assertion to make sure Draggables are safe to add
  - Add utility methods to the Anchor class to make it more "enum like"
- - `PositionComponent` is no longer an abstract class
+ - Enable user-defined anchors
+ - Added `toImage` method for the `Sprite` class
  - Add more optional arguments for unified constructors of components
 
 ## 1.0.0-rc6